requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
